Criação do DAOPedido
Criação do DAOPedido, alteração de campos de codigo definidos como
inteiro para Long. Mudança do banco da data de int para Date pendente

+++ Pizzaria/src/com/br/mackenzie/pizzaria/model/javabeans/Pedido.java
@@ -15,7 +15,7 @@ import java.util.List;
  */
 public class Pedido implements Serializable {
 
-    private int codigo;
+    private long codigo;
     private Cliente cliente;
     private double precoTotal;
     private Date data;
@@ -31,11 +31,11 @@ public class Pedido implements Serializable {
         this.data = data;
     }
 
-    public int getCodigo() {
+    public long getCodigo() {
         return codigo;
     }
 
-    public void setCodigo(int codigo) {
+    public void setCodigo(long codigo) {
         this.codigo = codigo;
     }
